+++ src/rai_core/rai/tools/ros2/generic/topics.py
@@ -32,8 +32,7 @@ from sensor_msgs.msg import CompressedImage, Image
 
 from rai.communication.ros2 import ROS2ARIConnector, ROS2ARIMessage
 from rai.communication.ros2.api.conversion import ros2_message_to_dict
-from rai.messages.multimodal import MultimodalArtifact
-from rai.messages.utils import preprocess_image
+from rai.messages import MultimodalArtifact, preprocess_image
 from rai.tools.ros2.base import BaseROS2Tool, BaseROS2Toolkit
